Add check for empty feed file in lookUpFeed
When a pipeline is initialized, the s3 resource will create an empty
file as its first version. Adding a check in the lookUpFeed task so
we error early in that scenario versus later in the pipeline.

+++ bakery/src/tasks/look-up-feed.js
@@ -28,6 +28,10 @@ const task = (taskArgs) => {
           dedent`
           exec 2> >(tee book/stderr >&2)
           feed="s3-feed/${versionedFile}"
+          if [[ ! -s "$feed" ]]; then
+            echo "Feed is empty"
+            exit 1
+          fi
           echo -n "$(cat $feed | jq -r '.collection_id')" >book/collection_id
           echo -n "$(cat $feed | jq -r '.server')" >book/server
           echo -n "$(cat $feed | jq -r '.style')" >book/style
